Moving mongo jobstore into mongo package
